+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 ------------------
 
 - #2120 Support for Multiple component analysis
+- #2123 Add Sample Form: Save and Copy Action
 - #2119 Fix linked client contact user can not see existing samples
 - #2118 Customized Quickinstaller Configlet
 - #2117 Customized User/Groups Preferences in Site Configuration